Use symbolic links instead of hard links in runtime pack directories.
The hard links are causing issues with package tooling.
This changes to use symbolic links instead.

Besides fixing the issues with the package tooling,
symbolic links are easier to preserve throughout
the packaging process.
